refactor : 파일 이름 수정, 불필요한 import, 파일 삭제
