+++ routes/location.js
@@ -1,13 +1,13 @@
 const express = require("express");
-const {applyLocation, readLocations, readLocation} = require("../controller/location");
+const locationController = require("../controllers/locationController");
 const {checkAuthenticated} = require("../middlewares/authMiddleware");
 const router = express.Router();
 
-router.get("/",readLocations)
+router.get("/",locationController.readLocations)
 //당분간 로그인이 필요없는 장소등록으로 배포
 // router.post("/",checkAuthenticated,applyLocation);
-router.post("/",applyLocation)
-router.get("/:id",readLocation);
+router.post("/",locationController.applyLocation)
+router.get("/:id",locationController.readLocation);
 
 
 module.exports = router;
